feat: change type price in migration int to numeric

+++ migrations/1552186891805-create-book-copy.ts
@@ -23,7 +23,7 @@ export class CreateBookCopy1552186891805 implements MigrationInterface {
           },
           {
             name: 'price',
-            type: 'int',
+            type: 'numeric',
             isNullable: true
           },
           {
